+++ fpdf/line_break.py
@@ -20,6 +20,7 @@ HYPHEN = "\u002d"
 SPACE = " "
 NBSP = "\u00a0"
 NEWLINE = "\n"
+FORM_FEED = "\u000c"
 
 
 class Fragment:
@@ -321,6 +322,7 @@ class TextLine(NamedTuple):
     height: float
     max_width: float
     trailing_nl: bool = False
+    trailing_form_feed: bool = False
 
 
 class SpaceHint(NamedTuple):
@@ -460,7 +462,9 @@ class CurrentLine:
         self.number_of_spaces = break_hint.number_of_spaces
         self.width = break_hint.line_width
 
-    def manual_break(self, align: Align, trailing_nl: bool = False):
+    def manual_break(
+        self, align: Align, trailing_nl: bool = False, trailing_form_feed: bool = False
+    ):
         return TextLine(
             fragments=self.fragments,
             text_width=self.width,
@@ -469,6 +473,7 @@ class CurrentLine:
             height=self.height,
             max_width=self.max_width,
             trailing_nl=trailing_nl,
+            trailing_form_feed=trailing_form_feed,
         )
 
     def automatic_break_possible(self):
@@ -610,12 +615,14 @@ class MultiLineBreak:
             )
             first_char = False
 
-            if character == NEWLINE:
+            if character in (NEWLINE, FORM_FEED):
                 self.character_index += 1
                 if not current_line.fragments:
                     current_line.height = current_font_height * self.line_height
                 return current_line.manual_break(
-                    Align.L if self.align == Align.J else self.align, trailing_nl=True
+                    Align.L if self.align == Align.J else self.align,
+                    trailing_nl=character == NEWLINE,
+                    trailing_form_feed=character == FORM_FEED,
                 )
             if current_line.width + character_width > max_width:
                 if character == SPACE:  # must come first, always drop a current space.