+++ test/image/test_vector_image.py
@@ -9,13 +9,14 @@ from test.conftest import assert_pdf_equal
 
 
 HERE = Path(__file__).resolve().parent
+SVG_SRCDIR = HERE.parent / "svg/svg_sources"
 
 
 def test_svg_image(tmp_path):
     pdf = fpdf.FPDF()
     pdf.add_page()
     # This image has a 300x300 viewbox and width=100% and height=100%:
-    pdf.image(HERE / "../svg/svg_sources/SVG_logo.svg")
+    pdf.image(SVG_SRCDIR / "SVG_logo.svg")
     assert_pdf_equal(pdf, HERE / "svg_image.pdf", tmp_path)
 
 
@@ -23,7 +24,7 @@ def test_svg_image_fixed_dimensions(tmp_path):
     pdf = fpdf.FPDF()
     pdf.add_page()
     # This image has a 300x300 viewbox and width=100 and height=100:
-    img = pdf.image(HERE / "../svg/svg_sources/SVG_logo_fixed_dimensions.svg")
+    img = pdf.image(SVG_SRCDIR / "SVG_logo_fixed_dimensions.svg")
     assert img["rendered_width"] == 100
     assert img["rendered_height"] == 100
     assert_pdf_equal(pdf, HERE / "svg_image_fixed_dimensions.pdf", tmp_path)
@@ -33,7 +34,7 @@ def test_svg_image_no_dimensions(tmp_path):
     pdf = fpdf.FPDF()
     pdf.add_page()
     # This image has a 300x300 viewbox but no width/height:
-    pdf.image(HERE / "../svg/svg_sources/SVG_logo_no_dimensions.svg")
+    pdf.image(SVG_SRCDIR / "SVG_logo_no_dimensions.svg")
     assert_pdf_equal(pdf, HERE / "svg_image_no_dimensions.pdf", tmp_path)
 
 
@@ -41,7 +42,7 @@ def test_svg_image_no_viewbox(tmp_path):
     pdf = fpdf.FPDF()
     pdf.add_page()
     # This image has no viewbox and width=100 and height=200:
-    img = pdf.image(HERE / "../svg/svg_sources/simple_rect_no_viewbox.svg")
+    img = pdf.image(SVG_SRCDIR / "simple_rect_no_viewbox.svg")
     assert img["rendered_width"] == 100
     assert img["rendered_height"] == 200
     assert_pdf_equal(pdf, HERE / "svg_image_no_viewbox.pdf", tmp_path)
@@ -51,7 +52,7 @@ def test_svg_image_with_custom_width(tmp_path):
     pdf = fpdf.FPDF()
     pdf.add_page()
     # This image has a 300x300 viewbox and width=100% and height=100%:
-    img = pdf.image(HERE / "../svg/svg_sources/SVG_logo.svg", w=60)
+    img = pdf.image(SVG_SRCDIR / "SVG_logo.svg", w=60)
     assert img["rendered_width"] == 60
     assert_pdf_equal(pdf, HERE / "svg_image_with_custom_width.pdf", tmp_path)
 
@@ -60,7 +61,7 @@ def test_svg_image_with_custom_width_and_no_dimensions(tmp_path):
     pdf = fpdf.FPDF()
     pdf.add_page()
     # This image has a 300x300 viewbox but no width/height:
-    img = pdf.image(HERE / "../svg/svg_sources/SVG_logo_no_dimensions.svg", w=60)
+    img = pdf.image(SVG_SRCDIR / "SVG_logo_no_dimensions.svg", w=60)
     assert img["rendered_width"] == 60
     assert_pdf_equal(
         pdf, HERE / "svg_image_with_custom_width_and_no_dimensions.pdf", tmp_path
@@ -71,7 +72,7 @@ def test_svg_image_with_custom_width_and_no_viewbox(tmp_path):
     pdf = fpdf.FPDF()
     pdf.add_page()
     # This image has no viewbox and width=100 and height=200:
-    img = pdf.image(HERE / "../svg/svg_sources/simple_rect_no_viewbox.svg", w=60)
+    img = pdf.image(SVG_SRCDIR / "simple_rect_no_viewbox.svg", w=60)
     assert img["rendered_width"] == 60
     assert_pdf_equal(
         pdf, HERE / "svg_image_with_custom_width_and_no_viewbox.pdf", tmp_path
@@ -82,7 +83,7 @@ def test_svg_image_with_no_dimensions_and_custom_width(tmp_path):
     pdf = fpdf.FPDF()
     pdf.add_page()
     # This image has a 300x300 viewbox but no width/height:
-    img = pdf.image(HERE / "../svg/svg_sources/SVG_logo_no_dimensions.svg", w=60)
+    img = pdf.image(SVG_SRCDIR / "SVG_logo_no_dimensions.svg", w=60)
     assert img["rendered_width"] == 60
     assert_pdf_equal(
         pdf, HERE / "svg_image_with_no_dimensions_and_custom_width.pdf", tmp_path
@@ -93,9 +94,7 @@ def test_svg_image_with_custom_size(tmp_path):
     pdf = fpdf.FPDF()
     pdf.add_page()
     # This image has a 300x300 viewbox but no width/height:
-    pdf.image(
-        HERE / "../svg/svg_sources/SVG_logo_no_dimensions.svg", x=50, y=50, w=30, h=60
-    )
+    pdf.image(SVG_SRCDIR / "SVG_logo_no_dimensions.svg", x=50, y=50, w=30, h=60)
     pdf.rect(x=50, y=50, w=30, h=60)  # Displays the bounding box
     assert_pdf_equal(pdf, HERE / "svg_image_with_custom_size.pdf", tmp_path)
 
@@ -104,9 +103,7 @@ def test_svg_image_with_custom_size_and_no_viewbox(tmp_path):
     pdf = fpdf.FPDF()
     pdf.add_page()
     # This image has no viewbox and width=100 and height=200:
-    img = pdf.image(
-        HERE / "../svg/svg_sources/simple_rect_no_viewbox.svg", x=50, y=50, w=30, h=60
-    )
+    img = pdf.image(SVG_SRCDIR / "simple_rect_no_viewbox.svg", x=50, y=50, w=30, h=60)
     assert img["rendered_width"] == 30
     assert img["rendered_height"] == 60
     pdf.rect(x=50, y=50, w=30, h=60)  # Displaying the bounding box
@@ -119,12 +116,10 @@ def test_svg_image_no_viewbox_nor_width_and_height():
     pdf = fpdf.FPDF()
     pdf.add_page()
     with pytest.raises(ValueError):
-        pdf.image(
-            HERE / "../svg/svg_sources/simple_rect_no_viewbox_nor_width_and_height.svg"
-        )
+        pdf.image(SVG_SRCDIR / "simple_rect_no_viewbox_nor_width_and_height.svg")
     with pytest.raises(ValueError):
         pdf.image(
-            HERE / "../svg/svg_sources/simple_rect_no_viewbox_nor_width_and_height.svg",
+            SVG_SRCDIR / "simple_rect_no_viewbox_nor_width_and_height.svg",
             w=60,
         )
 
@@ -192,3 +187,46 @@ def test_svg_image_billion_laughs():
                 b"<lolz>&lol9;</lolz>"
             )
         )
+
+
+def test_svg_image_fit_rect(tmp_path):
+    """
+    Scale the image to fill the rectangle, keeping its aspect ratio,
+    and ensure it does overflow the rectangle width or height in the process.
+    """
+    pdf = fpdf.FPDF()
+    pdf.add_page()
+    test_file = SVG_SRCDIR / "SVG_logo.svg"
+
+    rect1 = 30, 30, 60, 100
+    pdf.rect(*rect1)
+    pdf.image(test_file, *rect1, keep_aspect_ratio=True)
+
+    rect2 = 100, 30, 100, 60
+    pdf.rect(*rect2)
+    pdf.image(test_file, *rect2, keep_aspect_ratio=True)
+
+    assert_pdf_equal(pdf, HERE / "svg_image_fit_rect.pdf", tmp_path)
+
+
+IMG_DESCRIPTION = "The Mighty SVG Logo"
+
+
+def test_svg_image_alt_text_title(tmp_path):
+    test_file = SVG_SRCDIR / "SVG_logo.svg"
+    pdf = fpdf.FPDF()
+    pdf.add_page()
+    pdf.image(test_file, alt_text=IMG_DESCRIPTION, w=50, h=50)
+    pdf.image(test_file, title=IMG_DESCRIPTION, w=50, h=50)
+    pdf.image(test_file, alt_text=IMG_DESCRIPTION, title=IMG_DESCRIPTION, w=50, h=50)
+    assert_pdf_equal(pdf, HERE / "svg_image_alt_text_title.pdf", tmp_path)
+
+
+def test_svg_image_alt_text_two_pages(tmp_path):
+    test_file = SVG_SRCDIR / "SVG_logo.svg"
+    pdf = fpdf.FPDF()
+    pdf.add_page()
+    pdf.image(test_file, alt_text=IMG_DESCRIPTION, w=50, h=50)
+    pdf.add_page()
+    pdf.image(test_file, alt_text=IMG_DESCRIPTION, w=50, h=50)
+    assert_pdf_equal(pdf, HERE / "svg_image_alt_text_two_pages.pdf", tmp_path)